+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add Metrics API (#3791): Read our [docs](https://docs.sentry.io/platforms/apple/metrics/) to learn
 more about how to use the Metrics API.
 - Pre-main profiling data is now attached to the app start transaction (#3736)
+- Release framework without UIKit/AppKit (#3793)
 
 ### Fixes